generalized studio work and added info text for inheritance
added what we did in studio but made it so that when you select a user and click on any file, the permissions panel shows that user's permissions for the selected file. Can be done for any selected file and user.

Added some text in the popup box to explain that grey checkmarks indicate inherited permissions and blue checkmarks indicate that the file is the source of the rule.